[ENH] Delete task from worker if the data has been successfully saved in...

See merge request indu/rt-ramsis!117

+++ RAMSIS/cli/model.py
@@ -1,4 +1,7 @@
 import typer
+from datetime import datetime
+from rich import print
+from rich.table import Table
 from marshmallow import EXCLUDE
 from RAMSIS.db import db_url, session_handler, init_db
 from ramsis.datamodel import ModelConfig
@@ -13,30 +16,75 @@ app = typer.Typer()
 
 
 @app.command()
-def delete(
-        model_name: str,
-        force: bool = typer.Option(
-            False, help="Force the deletes without asking")):
+def ls(help="Outputs list of models"):
+    with session_handler(db_url) as session:
+        model_configs = session.execute(
+            select(ModelConfig)).scalars().all()
+        for model in model_configs:
+            table = Table(show_footer=False,
+                          title=f"Model Config {model.name}",
+                          title_justify="left")
+            table.add_column("attribute")
+            table.add_column("value")
+            for attr in ModelConfig.__table__.columns:
+                table.add_row(str(attr.name), str(getattr(model, attr.name)))
+
+            print(table)
+
+
+@app.command()
+def disable(
+        model_name: str):
     with session_handler(db_url) as session:
         model_config = session.execute(
             select(ModelConfig).filter_by(
                 name=model_name)).\
             scalar_one_or_none()
         if not model_config:
-            typer.echo("Model does not exist")
+            print("Model does not exist")
             raise typer.Exit()
-        if model_config.runs:
-            typer.echo("Model config is associated with model runs, "
-                       "cannot delete. You can disable instead.")
-            raise typer.Exit()
-            if not force:
-                delete = typer.confirm("Are you sure you want to delete the  "
-                                       f"model with name: {model_name}?")
-                if not delete:
-                    typer.echo("Not deleting")
-                    raise typer.Abort()
-        session.delete(model_config)
+        model_config.enabled = False
         session.commit()
+
+
+@app.command()
+def enable(
+        model_name: str):
+    with session_handler(db_url) as session:
+        model_config = session.execute(
+            select(ModelConfig).filter_by(
+                name=model_name)).\
+            scalar_one_or_none()
+        if not model_config:
+            print("Model does not exist")
+            raise typer.Exit()
+        model_config.enabled = True
+        session.commit()
+
+
+@app.command()
+def archive_all(
+        force: bool = typer.Option(
+            False, help="Force the deletes without asking")):
+    """Appends timestamp to the name of all models. For use
+    in cases when you change the model config/code and want to
+    archive the results.
+    """
+    with session_handler(db_url) as session:
+        model_configs = session.execute(
+            select(ModelConfig)).\
+            scalars().all()
+        if not model_configs:
+            print("Model does not exist")
+            raise typer.Exit()
+        for model in model_configs:
+            if 'ARCHIVED' not in model.name:
+                model.name = (f"{model.name}-ARCHIVED-"
+                              f"{datetime.now().strftime('%Y-%m-%d')}")
+                model.enabled = False
+        session.commit()
+        print("All models successfully archived: "
+              f"{[model.name for model in model_configs]}")
 
 
 @app.command()
@@ -51,7 +99,7 @@ def load(
     if success:
         pass
     else:
-        typer.echo(f"Error, db could not be initialized: {success}")
+        print(f"Error, db could not be initialized: {success}")
         raise typer.Exit()
     with session_handler(db_url) as session:
         with open(model_config, "r") as model_read:
@@ -64,16 +112,16 @@ def load(
                 scalar_one_or_none()
 
             if existing_config:
-                typer.echo(f"Model config exists for {existing_config.name}.")
+                print(f"Model config exists for {existing_config.name}.")
                 # Check if there are already completed forecasts
                 # associated before allowing modification.
-                runs_exist = True if existing_config.runs else False
-                if runs_exist:
+                print(existing_config.runs, "existing runs")
+                if existing_config.runs:
                     session.rollback()
-                    typer.echo("Model runs already exist for this config"
-                               " Please upload the config with a new name.")
-                    typer.Exit(code=1)
-                typer.echo("deleting existing config")
+                    print("Model runs already exist for this config"
+                          " Please upload the config with a new name.")
+                    raise typer.Exit(code=1)
+                print("deleting existing config")
                 session.delete(existing_config)
                 new_model_config = ModelConfigurationSchema(
                     unknown=EXCLUDE, context={"session": session}).load(config)
@@ -81,10 +129,10 @@ def load(
             else:
                 new_model_config = ModelConfigurationSchema(
                     unknown=EXCLUDE, context={"session": session}).load(config)
-                typer.echo("Model config is being added for "
-                           f"{new_model_config.name}")
+                print("Model config is being added for "
+                      f"{new_model_config.name}")
 
             session.add(new_model_config)
             session.commit()
-            typer.echo("A model has been configured with the name: "
-                       f"{new_model_config.name}, id: {new_model_config.id}")
+            print("A model has been configured with the name: "
+                  f"{new_model_config.name}, id: {new_model_config.id}")
